accumulated PRs

+++ bbb-common-web/src/main/java/org/bigbluebutton/presentation/imp/SvgImageCreatorImp.java
@@ -29,8 +29,10 @@ public class SvgImageCreatorImp implements SvgImageCreator {
     private SwfSlidesGenerationProgressNotifier notifier;
     private long imageTagThreshold;
     private long pathsThreshold;
-    private String convTimeout = "60s";
-    private int WAIT_FOR_SEC = 60;
+    private int convPdfToSvgTimeout = 60;
+    private int svgResolutionPpi = 300;
+    private boolean forceRasterizeSlides = false;
+    private int pngWidthRasterizedSlides = 2048;
 	private String BLANK_SVG;
 
     @Override
@@ -64,18 +66,22 @@ public class SvgImageCreatorImp implements SvgImageCreator {
             dest = imagePresentationDir.getAbsolutePath() + File.separator + "slide-1.pdf";
 
             NuProcessBuilder convertImgToSvg = new NuProcessBuilder(
-                    Arrays.asList("timeout", convTimeout, "convert", source, "-auto-orient", dest));
+                    Arrays.asList("timeout", convPdfToSvgTimeout + "s", "convert", source, "-auto-orient", dest));
 
             Png2SvgConversionHandler pHandler = new Png2SvgConversionHandler();
             convertImgToSvg.setProcessListener(pHandler);
 
             NuProcess process = convertImgToSvg.start();
             try {
-                process.waitFor(WAIT_FOR_SEC, TimeUnit.SECONDS);
+                process.waitFor(convPdfToSvgTimeout + 1, TimeUnit.SECONDS);
                 done = true;
             } catch (InterruptedException e) {
                 done = false;
                 log.error("InterruptedException while converting to SVG {}", dest, e);
+            }
+
+            if(pHandler.isCommandTimeout()) {
+                log.error("Command execution (convertImgToSvg) exceeded the {} secs timeout for {} page {}.", convPdfToSvgTimeout, pres.getName(), page);
             }
 
             // Use the intermediate PDF file as source
@@ -89,26 +95,37 @@ public class SvgImageCreatorImp implements SvgImageCreator {
 
         File destsvg = new File(imagePresentationDir.getAbsolutePath() + File.separatorChar + "slide" + page + ".svg");
 
-        NuProcessBuilder convertPdfToSvg = createConversionProcess("-svg", page, source, destsvg.getAbsolutePath(),
+        SvgConversionHandler pHandler = new SvgConversionHandler();
+
+        if(this.forceRasterizeSlides == false) {
+            NuProcessBuilder convertPdfToSvg = createConversionProcess("-svg", page, source, destsvg.getAbsolutePath(),
                     true);
 
-        SvgConversionHandler pHandler = new SvgConversionHandler();
-        convertPdfToSvg.setProcessListener(pHandler);
+            convertPdfToSvg.setProcessListener(pHandler);
 
-        NuProcess process = convertPdfToSvg.start();
-        try {
-            process.waitFor(WAIT_FOR_SEC, TimeUnit.SECONDS);
-            done = true;
-        } catch (InterruptedException e) {
-            log.error("Interrupted Exception while generating SVG slides {}", pres.getName(), e);
+            NuProcess process = convertPdfToSvg.start();
+            try {
+                process.waitFor(convPdfToSvgTimeout + 1, TimeUnit.SECONDS);
+                done = true;
+            } catch (InterruptedException e) {
+                log.error("Interrupted Exception while generating SVG slides {}", pres.getName(), e);
+            }
+
+            if(pHandler.isCommandTimeout()) {
+                log.error("Command execution (convertPdfToSvg) exceeded the {} secs timeout for {} page {}.", convPdfToSvgTimeout, pres.getName(), page);
+            }
+
+            if (!done) {
+                return done;
+            }
         }
 
-        if (!done) {
-            return done;
-        }
 
-        if (destsvg.length() == 0 || pHandler.numberOfImageTags() > imageTagThreshold
-                || pHandler.numberOfPaths() > pathsThreshold) {
+        if (destsvg.length() == 0 ||
+                pHandler.numberOfImageTags() > imageTagThreshold ||
+                pHandler.numberOfPaths() > pathsThreshold ||
+                this.forceRasterizeSlides) {
+
             // We need t delete the destination file as we are starting a
             // new conversion process
             if (destsvg.exists()) {
@@ -117,21 +134,24 @@ public class SvgImageCreatorImp implements SvgImageCreator {
 
             done = false;
 
-            Map<String, Object> logData = new HashMap<String, Object>();
-            logData.put("meetingId", pres.getMeetingId());
-            logData.put("presId", pres.getId());
-            logData.put("filename", pres.getName());
-            logData.put("page", page);
-            logData.put("convertSuccess", pHandler.isCommandSuccessful());
-            logData.put("fileExists", destsvg.exists());
-            logData.put("numberOfImages", pHandler.numberOfImageTags());
-            logData.put("numberOfPaths", pHandler.numberOfPaths());
-            logData.put("logCode", "potential_problem_with_svg");
-            logData.put("message", "Potential problem with generated SVG");
-            Gson gson = new Gson();
-            String logStr = gson.toJson(logData);
+            if(!this.forceRasterizeSlides) {
+                Map<String, Object> logData = new HashMap<String, Object>();
+                logData.put("meetingId", pres.getMeetingId());
+                logData.put("presId", pres.getId());
+                logData.put("filename", pres.getName());
+                logData.put("page", page);
+                logData.put("convertSuccess", pHandler.isCommandSuccessful());
+                logData.put("fileExists", destsvg.exists());
+                logData.put("numberOfImages", pHandler.numberOfImageTags());
+                logData.put("numberOfPaths", pHandler.numberOfPaths());
+                logData.put("logCode", "potential_problem_with_svg");
+                logData.put("message", "Potential problem with generated SVG");
+                Gson gson = new Gson();
+                String logStr = gson.toJson(logData);
 
-            log.warn(" --analytics-- data={}", logStr);
+                log.warn(" --analytics-- data={}", logStr);
+            }
+
 
             File tempPng = null;
             String basePresentationame = UUID.randomUUID().toString();
@@ -140,14 +160,15 @@ public class SvgImageCreatorImp implements SvgImageCreator {
             } catch (IOException ioException) {
                 // We should never fall into this if the server is correctly
                 // configured
+                Map<String, Object> logData = new HashMap<String, Object>();
                 logData = new HashMap<String, Object>();
                 logData.put("meetingId", pres.getMeetingId());
                 logData.put("presId", pres.getId());
                 logData.put("filename", pres.getName());
                 logData.put("logCode", "problem_with_creating_svg");
                 logData.put("message", "Unable to create temporary files");
-                gson = new Gson();
-                logStr = gson.toJson(logData);
+                Gson gson = new Gson();
+                String logStr = gson.toJson(logData);
                 log.error(" --analytics-- data={}", logStr, ioException);
             }
 
@@ -159,46 +180,65 @@ public class SvgImageCreatorImp implements SvgImageCreator {
             convertPdfToPng.setProcessListener(pngHandler);
             NuProcess pngProcess = convertPdfToPng.start();
             try {
-                pngProcess.waitFor(WAIT_FOR_SEC, TimeUnit.SECONDS);
+                pngProcess.waitFor(convPdfToSvgTimeout + 1, TimeUnit.SECONDS);
             } catch (InterruptedException e) {
                 log.error("Interrupted Exception while generating PNG image {}", pres.getName(), e);
             }
 
-            // Step 2: Convert a PNG image to SVG
-            NuProcessBuilder convertPngToSvg = new NuProcessBuilder(Arrays.asList("timeout", convTimeout, "convert",
-                        tempPng.getAbsolutePath(), destsvg.getAbsolutePath()));
-
-            Png2SvgConversionHandler svgHandler = new Png2SvgConversionHandler();
-            convertPngToSvg.setProcessListener(svgHandler);
-            NuProcess svgProcess = convertPngToSvg.start();
-            try {
-                svgProcess.waitFor(WAIT_FOR_SEC, TimeUnit.SECONDS);
-            } catch (InterruptedException e) {
-                log.error("Interrupted Exception while generating SVG image {}", pres.getName(), e);
+            if(pngHandler.isCommandTimeout()) {
+                log.error("Command execution (convertPdfToPng) exceeded the {} secs timeout for {} page {}.", convPdfToSvgTimeout, pres.getName(), page);
             }
 
-            done = svgHandler.isCommandSuccessful();
+            if(tempPng.length() > 0) {
+                // Step 2: Convert a PNG image to SVG
+                NuProcessBuilder convertPngToSvg = new NuProcessBuilder(Arrays.asList("timeout", convPdfToSvgTimeout + "s", "convert",
+                            tempPng.getAbsolutePath(), destsvg.getAbsolutePath()));
+
+                Png2SvgConversionHandler svgHandler = new Png2SvgConversionHandler();
+                convertPngToSvg.setProcessListener(svgHandler);
+                NuProcess svgProcess = convertPngToSvg.start();
+                try {
+                    svgProcess.waitFor(convPdfToSvgTimeout + 1, TimeUnit.SECONDS);
+                } catch (InterruptedException e) {
+                    log.error("Interrupted Exception while generating SVG image {}", pres.getName(), e);
+                }
+
+                if(svgHandler.isCommandTimeout()) {
+                    log.error("Command execution (convertPngToSvg) exceeded the {} secs timeout for {} page {}.", convPdfToSvgTimeout, pres.getName(), page);
+                }
+
+                done = svgHandler.isCommandSuccessful();
+
+                if(destsvg.length() > 0) {
+                    // Step 3: Add SVG namespace to the destionation file
+                    // Check : https://phabricator.wikimedia.org/T43174
+                    NuProcessBuilder addNameSpaceToSVG = new NuProcessBuilder(Arrays.asList("timeout", convPdfToSvgTimeout + "s",
+                            "/bin/sh", "-c",
+                            "sed -i "
+                                    + "'4s|>| xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" version=\"1.2\">|' "
+                                    + destsvg.getAbsolutePath()));
+
+                    AddNamespaceToSvgHandler namespaceHandler = new AddNamespaceToSvgHandler();
+                    addNameSpaceToSVG.setProcessListener(namespaceHandler);
+                    NuProcess namespaceProcess = addNameSpaceToSVG.start();
+                    try {
+                        namespaceProcess.waitFor(convPdfToSvgTimeout + 1, TimeUnit.SECONDS);
+                    } catch (InterruptedException e) {
+                        log.error("Interrupted Exception while adding SVG namespace {}", pres.getName(), e);
+                    }
+
+                    if (namespaceHandler.isCommandTimeout()) {
+                        log.error("Command execution (addNameSpaceToSVG) exceeded the {} secs timeout for {} page {}.", convPdfToSvgTimeout, pres.getName(), page);
+                    }
+                }
+            }
 
             // Delete the temporary PNG after finishing the image conversion
-            tempPng.delete();
-
-            // Step 3: Add SVG namespace to the destionation file
-            // Check : https://phabricator.wikimedia.org/T43174
-            NuProcessBuilder addNameSpaceToSVG = new NuProcessBuilder(Arrays.asList("timeout", convTimeout,
-                        "/bin/sh", "-c",
-                        "sed -i "
-                                + "'4s|>| xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" version=\"1.2\">|' "
-                                + destsvg.getAbsolutePath()));
-
-            AddNamespaceToSvgHandler namespaceHandler = new AddNamespaceToSvgHandler();
-            addNameSpaceToSVG.setProcessListener(namespaceHandler);
-            NuProcess namespaceProcess = addNameSpaceToSVG.start();
-            try {
-                namespaceProcess.waitFor(WAIT_FOR_SEC, TimeUnit.SECONDS);
-            } catch (InterruptedException e) {
-                log.error("Interrupted Exception while adding SVG namespace {}", pres.getName(), e);
+            if(tempPng.exists()) {
+                tempPng.delete();
             }
         }
+
 
         long endConv = System.currentTimeMillis();
 
@@ -226,13 +266,20 @@ public class SvgImageCreatorImp implements SvgImageCreator {
 
     private NuProcessBuilder createConversionProcess(String format, int page, String source, String destFile,
             boolean analyze) {
-        String rawCommand = "pdftocairo -r " + (analyze ? " 300 " : " 150 ") + format + (analyze ? "" : " -singlefile")
-                + " -q -f " + String.valueOf(page) + " -l " + String.valueOf(page) + " " + source + " " + destFile;
+        String rawCommand = "pdftocairo -r " + this.svgResolutionPpi + " " + format + (analyze ? "" : " -singlefile");
+
+        //Resize png resolution to avoid too large files
+        if(format.equals("-png") && this.pngWidthRasterizedSlides != 0) {
+            rawCommand += " -scale-to-x " + this.pngWidthRasterizedSlides + " -scale-to-y -1 ";
+        }
+
+        rawCommand  += " -q -f " + String.valueOf(page) + " -l " + String.valueOf(page) + " " + source + " " + destFile;
         if (analyze) {
             rawCommand += " && cat " + destFile;
             rawCommand += " | egrep 'data:image/png;base64|<path' | sed 's/  / /g' | cut -d' ' -f 1 | sort | uniq -cw 2";
         }
-        return new NuProcessBuilder(Arrays.asList("timeout", convTimeout, "/bin/sh", "-c", rawCommand));
+
+        return new NuProcessBuilder(Arrays.asList("timeout", convPdfToSvgTimeout + "s", "/bin/sh", "-c", rawCommand));
     }
 
     private File determineSvgImagesDirectory(File presentationFile) {
@@ -277,5 +324,21 @@ public class SvgImageCreatorImp implements SvgImageCreator {
     public void setSwfSlidesGenerationProgressNotifier(
         SwfSlidesGenerationProgressNotifier notifier) {
       this.notifier = notifier;
+    }
+
+    public void setConvPdfToSvgTimeout(int convPdfToSvgTimeout) {
+        this.convPdfToSvgTimeout = convPdfToSvgTimeout;
+    }
+
+    public void setSvgResolutionPpi(int svgResolutionPpi) {
+        this.svgResolutionPpi = svgResolutionPpi;
+    }
+
+    public void setForceRasterizeSlides(boolean forceRasterizeSlides) {
+        this.forceRasterizeSlides = forceRasterizeSlides;
+    }
+
+    public void setPngWidthRasterizedSlides(int pngWidthRasterizedSlides) {
+        this.pngWidthRasterizedSlides = pngWidthRasterizedSlides;
     }
 }
